requirements.txt
opencv-python
numpy
mediapipe
requests
Pillow
google-generativeai
